refactor: externalized GetService & unit tests
